feat: add route for getting partial task by id (#9)
* feat: add route for getting partial task by id

* chore(lint): fix lint issue

* chore: add TODO for custom errors

* test(integration): add test that return status code OK

* test: finish integration tests and start unit tests

* test(unit): added some more unit tests WIP

* test(unit): finished unit tests

* chore: fix lint errors

* chore: fixes according to requests

* chore: fix requested changes

+++ src/partialTask/models/partialTaskManager.ts
@@ -2,10 +2,17 @@ import { DeleteResult } from 'typeorm';
 import { inject, injectable } from 'tsyringe';
 import { SearchOrder, Services } from '../../common/constants';
 import convertTaskEntityToResponse from '../../common/utils/convertTaskEntityToResponse';
-import { ILogger, IPartialTaskCreate, IPartialTaskParams, IPartialTaskResponse, IPartialTaskStatusUpdate } from '../../common/interfaces';
+import {
+  IDiscreteTaskParams,
+  ILogger,
+  IPartialTaskCreate,
+  IPartialTaskParams,
+  IPartialTaskResponse,
+  IPartialTaskStatusUpdate,
+} from '../../common/interfaces';
 import { ConnectionManager } from '../../DAL/connectionManager';
 import { PartialTaskRepository } from '../../DAL/repositories/partialTaskRepository';
-import { DiscreteTaskEntity } from '../../DAL/entity/discreteTask';
+import { DiscreteTaskManager } from '../../discreteTask/models/discreteTaskManager';
 
 @injectable()
 export class PartialTaskManager {
@@ -13,12 +20,41 @@ export class PartialTaskManager {
 
   public constructor(@inject(Services.LOGGER) private readonly logger: ILogger, private readonly connectionManager: ConnectionManager) {}
 
-  public async getPartialTasksByDiscrete(discrete: DiscreteTaskEntity, order: SearchOrder): Promise<IPartialTaskResponse[]> {
+  public async getPartialTask(params: IPartialTaskParams): Promise<IPartialTaskResponse> {
     const repository = await this.getRepository();
-    const records = await repository.getAll(discrete, order);
 
+    const exists = await repository.exists(params);
+    // Check if discrete already exists
+    if (!exists) {
+      // TODO: throw custom error
+      throw new Error('No such partial task');
+    }
+
+    const record = await repository.get(params);
+    if (!record) {
+      // TODO: throw custom error
+      throw new Error('Failed to get partial task');
+    }
+
+    const model = convertTaskEntityToResponse(record);
+    return model;
+  }
+
+  public async getPartialTasksByDiscrete(discreteParams: IDiscreteTaskParams, order: SearchOrder): Promise<IPartialTaskResponse[]> {
+    const repository = await this.getRepository();
+
+    const discreteManager = new DiscreteTaskManager(this.logger, this.connectionManager);
+    const exists = await discreteManager.exists(discreteParams);
+
+    if (!exists) {
+      // TODO: throw custom error
+      throw new Error('No such discrete');
+    }
+
+    const records = await repository.getAll(discreteParams, order);
     if (!records) {
-      return Promise.reject();
+      // TODO: throw custom error
+      throw new Error('Failed to get all partial tasks by discrete');
     }
 
     const model = records.map((task) => convertTaskEntityToResponse(task));
@@ -32,15 +68,16 @@ export class PartialTaskManager {
     };
 
     const exists = await repository.exists(task);
-    // Check if discrete already exists
+    // Check if partial task already exists
     if (!exists) {
-      return Promise.reject();
+      // TODO: throw custom error
+      throw new Error('No such partial task');
     }
 
     const record = await repository.updatePartialTask(params);
-
     if (!record) {
-      return Promise.reject();
+      // TODO: throw custom error
+      throw new Error('Failed to update partial task');
     }
 
     const model = convertTaskEntityToResponse(record);
@@ -49,10 +86,11 @@ export class PartialTaskManager {
 
   public async createResource(resource: IPartialTaskCreate): Promise<IPartialTaskResponse> {
     const repository = await this.getRepository();
-    const record = await repository.createPartialTask(resource);
 
+    const record = await repository.createPartialTask(resource);
     if (!record) {
-      return Promise.reject();
+      // TODO: throw custom error
+      throw new Error('Error creating partial task');
     }
 
     const model = convertTaskEntityToResponse(record);
@@ -61,6 +99,14 @@ export class PartialTaskManager {
 
   public async deleteResource(task: IPartialTaskParams): Promise<DeleteResult> {
     const repository = await this.getRepository();
+
+    // Check if discrete exists
+    const partial = await repository.get(task);
+    if (!partial) {
+      // TODO: throw custom error
+      throw new Error('Partial task does not exist');
+    }
+
     const res = await repository.deletePartialTask(task);
     return res;
   }
